This is proposal of a way to deal with errors and states in the app.
The idea is to avoid using the onError/onCompleted channel on the
BehaviourSubjects. Instead we manually call onNext on them and we use
two channels for data/events.

The event channel can then be hooked to error displays and loading
states while the data one is free to simply deal with data display.