Login implementation for Firebase w/ username and password
KazakAuth is a generic interface for logging in via different methods, right now
you can only log in with username and password (to be added: google at least).
FirebaseKazakAuth is the concrete implementation, proxying the login to the
Firebase SDK.
ScheduleObserver in KazakDataRepository was refactored into a generic Observer
(similar to what @dorvaryn has already merged into develop).