Download and put updated schedule in assets
This is a proposal to download and save the most updated schedule on
compile time. The `download{FLAVOR}{BUILD_TYPE}Schedule` takes care
of downloading the events schedule from the Web Service specified in the
`flavor`-related file.

This implementation forces the download in the build/intermediates/assets
directory right after all the other assets have been merged into that same dir.
We could also avoid this automatism and still launch the task explicitly when
we want to update the schedule from the Web Service (at that point, it'd be
better to save the updated JSON in the src root folder, not the build one).

This PR also solves a couple of issues already solved in #39 (still pending).